% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ams.hext.woN}
\alias{ams.hext.woN}
\title{Analyse AMS data using Linear Perturbation Analysis method proposed by Hext, improved version}
\usage{
ams.hext.woN(measures, setup, alpha = 0.95)
}
\arguments{
\item{measures}{Object with class \code{ams.measures}.}

\item{setup}{Object with class \code{amssetup}.}

\item{alpha}{Number between 0 and 1. Confidence level.}
}
\description{
Analyse AMS data using Linear Perturbation Analysis method proposed by Hext, improved version
}
\examples{
# Sample AMS measures object
measures <- as.ams.measures(sample_measures)

# Experimental setup object
setup <- ams.setup()

# Estimate AMS parameters and confidence intervals with 0.99 confidence
# level
ams.hext.woN(measures, setup, alpha = 0.99)
}

