% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ams.setup}
\alias{ams.setup}
\title{Constructs a AMS experiment setup}
\usage{
ams.setup(setup.matrix = .__mat_D_default)
}
\arguments{
\item{setup.matrix}{Numeric matrix.}
}
\description{
Constructs a AMS experiment setup
}
\examples{
setup <- ams.setup()
class(setup)
}

