% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ams.sus_tensor}
\alias{ams.sus_tensor}
\title{ams.sus_tensor takes a ams.measures object and its setup and returns the
best estimate for the susceptibility tensor as a 3 times 3 matrix.}
\usage{
ams.sus_tensor(measures, setup)
}
\arguments{
\item{measures}{An object of class \code{ams.measures}.}

\item{setup}{An object of class \code{ams.setup}}
}
\description{
ams.sus_tensor takes a ams.measures object and its setup and returns the
best estimate for the susceptibility tensor as a 3 times 3 matrix.
}
\examples{
# Experimental setup object
setup <- ams.setup()
# Sample AMS measures object
measures <- as.ams.measures(sample_measures)
# Least squares estimation of the susceptibility tensor
ams.sus_tensor(measures, setup)
}

