% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AMSsetup}
\alias{AMSsetup}
\title{Constructs a AMS experiment setup}
\usage{
AMSsetup(setup.matrix = kSetupMatrixDefault)
}
\arguments{
\item{setup.matrix}{Numeric matrix.}
}
\description{
Constructs a AMS experiment setup
}
\examples{
setup <- AMSsetup()
class(setup)
}

