% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{amsstats}
\alias{amsstats}
\alias{amsstats-package}
\title{Statistics tools for AMS analysis}
\description{
This package provides some functions to analyse directional susceptibility
measurements to determine the susceptibility matrix and provide confidence
intervals.
}

