% Generated by roxygen2 (4.0.1): do not edit by hand
\name{anisotropytest}
\alias{anisotropytest}
\alias{anisotropytest.AMSanalysis}
\title{The results of the anisotropy test from an AMS analisys}
\usage{
anisotropytest(analysis)

\method{anisotropytest}{AMSanalysis}(analysis)
}
\arguments{
\item{analysis}{The AMS analysis object}
}
\description{
The anisotropy test specifies if the hipotesis of isotropy,
oblate or prolate ellipsoid can be rejected, that is, if
the different eigenvalues can be equal.
}

