% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.ams.measures}
\alias{as.ams.measures}
\title{Constructs an AMS measurements object.}
\usage{
as.ams.measures(data)
}
\arguments{
\item{data}{Data frame with at least three columns. Each row represents a
repetition of a measurement in one of the experimental setup positions.
The first columns must be integer and is interpreted as the number of
repetition of the measurement. The second columns, also
integer, is the order of the corresponding measurement in the experimental
setup. And the third column, numeric, the measured values.}
}
\description{
This method constructs an object with class \code{ams.measures}, which uses
a \code{data.frame} internally. It takes as argument a data frame with at
least three columns (list, data.frame, matrix,...).
}
\examples{
# Use sample dataframe
class(sample_measures)

# Convert data frame to \\code{ams.measures} object
ams_measure <- as.ams.measures(sample_measures)
class(sample_measures)
}

