% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BootstrapAnalyse}
\alias{BootstrapAnalyse}
\title{Analyse AMS data using Bootstrap method proposed by Constable and Tauxe}
\usage{
BootstrapAnalyse(measures, setup, alpha = 0.95, R = 1000,
  normalize = FALSE)
}
\arguments{
\item{measures}{Object with class \code{AMSmeasures}.}

\item{setup}{Object with class \code{amssetup}.}

\item{alpha}{Number between 0 and 1. Confidence level.}

\item{R}{Integer. Number of resamplings}

\item{normalize}{Boolean. Should the estimations of the AMS parameters
    for the resampled data be normalized before mixing them?}
}
\description{
This function takes a \code{AMSmeasures} object and returns the estimated
eigenvalues and their errors and the estimated eigenvectors with
their errors.
}
\examples{
# Sample AMS measures object
measures <- as.AMSmeasures(sample_measures)

# Experimental setup object
setup <- AMSsetup()

# Estimate AMS parameters and confidence intervals with 0.99 confidence
# level
BootstrapAnalyse(measures, setup, alpha = 0.99)

# Estimate AMS parameters and confidence intervals with 0.95 confidence
# level (default) and only 100 bootstrap resamples
BootstrapAnalyse(measures, setup, R = 100)
}

