% Generated by roxygen2 (4.0.1): do not edit by hand
\name{design}
\alias{design}
\alias{design.AMSsetup}
\title{Returns the design matrix for setup object}
\usage{
design(setup)

\method{design}{AMSsetup}(setup)
}
\arguments{
\item{setup}{Setup object}
}
\description{
Returns the design matrix for setup object
}

