% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eigenvalues}
\alias{eigenvalues}
\alias{eigenvalues.AMSanalysis}
\title{The eigenvalues confidence intervals from an AMS analisys}
\usage{
eigenvalues(analysis)

\method{eigenvalues}{AMSanalysis}(analysis)
}
\arguments{
\item{analysis}{The AMS analysis object}
}
\description{
Get the eigenvalues confidence intervals resulting from the AMS analysis.
}

