% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eigenvectors}
\alias{eigenvectors}
\alias{eigenvectors.AMSanalysis}
\title{The eigenvectors confidence regions from an AMS analisys}
\usage{
eigenvectors(analysis)

\method{eigenvectors}{AMSanalysis}(analysis)
}
\arguments{
\item{analysis}{The AMS analysis object}
}
\description{
Get the eigenvectors confidence regions, typically spherical ellipses,
resulting from the AMS analysis.
}

