% Generated by roxygen2 (4.0.1): do not edit by hand
\name{error_norm_dist_generator}
\alias{error_norm_dist_generator}
\title{Generates a normal noise generator for use in \code{fake_measurements} function.}
\usage{
error_norm_dist_generator(sigma)
}
\arguments{
\item{sigma}{Numeric. The standard deviation of the simulated normal noise.}
}
\description{
Generates a normal noise generator for use in \code{fake_measurements} function.
}
\examples{
# Normal error distribution function for synthetic AMS measures with
# standard debiation 0.1
error_dist <- error_norm_dist_generator(0.1)
}

