% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fake_measurements}
\alias{fake_measurements}
\title{Generates a synthetic AMS measurements object}
\usage{
fake_measurements(sus_matrix, n_measurements, error_dist, setup = ams.setup())
}
\arguments{
\item{sus_matrix}{Nueric 3x3 symetric matrix.}

\item{n_measurements}{Integer. The number of repetitions of the
simulated experimental protocol.}

\item{error_dist}{A function accepting two arguments and returning a matrix.
It must accept a numeric vector as the first argument and an integer
as the second one, and return an \code{n} times \code{m} numeric
matrix, being \code{n} the length of its first argument and \code{m}
the second one. This function is interpreted as a random generator
for the simulated AMS measures.}

\item{setup}{Object of class \code{ams.setup}.}
}
\description{
Generates a synthetic AMS measurements object
}
\examples{
# Reference ('real') susceptibility matrix
sus_matrix <- matrix(c(1.6,.5,.3, .5,1.3,.6, .3,.6,1.9), nrow = 3)
sus_matrix

# Simulated error distribution with standard deviation 0.5
error_dist <- error_norm_dist_generator(0.5)
# Experimental setup object
setup <- ams.setup()

# Fake AMS measurements object with 2 repetitions
fake_measurements(sus_matrix, 2, error_dist, setup)
}

