% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FakeMeasures}
\alias{FakeMeasures}
\title{Generates a synthetic AMS measurements object}
\usage{
FakeMeasures(suscept.matrix, nmeasures, error.dist, setup = AMSsetup())
}
\arguments{
\item{suscept.matrix}{Nueric 3x3 symetric matrix.}

\item{nmeasures}{Integer. The number of repetitions of the
simulated experimental protocol.}

\item{error.dist}{A function accepting two arguments and returning a matrix.
It must accept a numeric vector as the first argument and an integer
as the second one, and return an \code{n} times \code{m} numeric
matrix, being \code{n} the length of its first argument and \code{m}
the second one. This function is interpreted as a random generator
for the simulated AMS measures.}

\item{setup}{Object of class \code{AMSsetup}.}
}
\description{
Generates a synthetic AMS measurements object
}
\examples{
# Reference ('real') susceptibility matrix
suscept.matrix <- matrix(c(1.6,.5,.3, .5,1.3,.6, .3,.6,1.9), nrow = 3)
suscept.matrix

# Simulated error distribution with standard deviation 0.5
error.dist <- NormalErrorGenerator(0.5)
# Experimental setup object
setup <- AMSsetup()

# Fake AMS measurements object with 2 repetitions
FakeMeasures(suscept.matrix, 2, error.dist, setup)
}

