% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mean.AMSmeasures}
\alias{mean.AMSmeasures}
\title{Computes the mean of AMS measures.}
\usage{
\method{mean}{AMSmeasures}(x, ...)
}
\arguments{
\item{x}{An object with class \code{AMSmeasures}.}

\item{...}{Unused parameters}
}
\description{
Returns an \code{\link{AMSmeasures}} object with a single repetition and
the mean value for each measured position.
}

