% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.SpherEllipse}
\alias{plot.SpherEllipse}
\title{Plot spherical ellipse}
\usage{
\method{plot}{SpherEllipse}(x, npoints = 200, add = FALSE,
  line.col = "red", ...)
}
\arguments{
\item{x}{An SpherEllipse object}

\item{npoints}{The number of points to be plotted, default is 200}

\item{add}{Boolean indicating if the plot should be drawn over an existing
plot (\code{True}) or on a new one (\code{False}, default)}

\item{line.col}{The color of the ellipse, default is 'red'}

\item{...}{Other parameters to be passed to \code{\link{lambert.plot}}}
}
\description{
Plots an spherial ellipse in Lambert azimuthal (equal area)
projection. Only the north hemisphere
is represented, directions in the south hemisphere are reverted and
represented in the north pole.
}

