% Generated by roxygen2 (4.0.1): do not edit by hand
\name{poss}
\alias{poss}
\alias{poss.AMSmeasures}
\alias{values.AMSmeasures}
\title{Returns a vector of integers indicating which position in the experimental
setup the corresponding measured data corresponds to.}
\usage{
poss(measures)

\method{poss}{AMSmeasures}(measures)

\method{values}{AMSmeasures}(measures)
}
\arguments{
\item{measures}{AMS measures object}
}
\description{
Returns a vector of integers indicating which position in the experimental
setup the corresponding measured data corresponds to.
}

