% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reps}
\alias{reps}
\alias{reps.AMSmeasures}
\title{Returns a vector of integers indicating which measure or repetition
the corresponding measured data corresponds to.}
\usage{
reps(measures)

\method{reps}{AMSmeasures}(measures)
}
\arguments{
\item{measures}{AMS measures object}
}
\description{
Returns a vector of integers indicating which measure or repetition
the corresponding measured data corresponds to.
}

