% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{sample_measures}
\alias{sample_measures}
\title{Sample AMS measures}
\format{A data frame with 3 columns and 150 rows. Each row represents a
    measurement. The three columns represent repetition,
    position and value for each measurement.}
\source{
Synthetic data generated with \code{fake_measurements}.
}
\description{
Data frame with sample AMS experimental data, with 10
    repetitions and standard setup of 15 positions.
}

