% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_analysis}
\alias{sim_analysis}
\title{Performs simultations of AMS statistics analysis}
\usage{
sim_analysis(methods, sus_matrix, n_measurements, error_dist,
  setup = ams.setup(), ...)
}
\arguments{
\item{methods}{A character vector. The names of the AMS analysis methods
to be simulated.}

\item{sus_matrix}{Numeric 3x3 symmetric matrix.}

\item{n_measurements}{Integer. Number of repetitions of the simulated
experimental measures.}

\item{error_dist}{A function for simulating the experimental errors,
as in \code{fake_measurements}.}

\item{setup}{Object of class \code{ams.setup}.}

\item{...}{Extra arguments to be passed to the functions given in the
    \code{methods} argument.}
}
\description{
Performs simultations of AMS statistics analysis
}
\examples{
# Initial susceptibility tensor to be used as reference for the generation
# of simulated data
suscep0 <- matrix(c(1,0,0, 0,2,0, 0,0,3), nrow = 3)

sim_analysis(methods = c('ams.hext.woN', 'ams.constable'),
             sus_matrix = suscep0,
             n_measurements = 10,
             error_dist = error_norm_dist_generator(0.3),
             setup = ams.setup())
}

