% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SpherEllipse}
\alias{SpherEllipse}
\title{Constructs an spherical ellipse}
\usage{
SpherEllipse(centerDir, axis1Dir, axis2Dir, semiangle1, semiangle2)
}
\arguments{
\item{centerDir}{Numeric vector, representing longitud and latitude of the
center of the ellipse}

\item{axis1Dir}{Numeric vector, representing longitud and latitude of the
first semiaxis of the ellipse}

\item{axis2Dir}{Numeric vector, representing longitud and latitude of the
second semiaxis of the ellipse}

\item{semiangle1}{Numberic value, angular magnitud of the first semiaxis}

\item{semiangle2}{Numberic value, angular magnitud of the second semiaxis}
}
\description{
An ellipse in the sphere is described by a list object 'elip' with constituent
elements 'elip$center', 'elip$eta', 'elip$zeta', 'elip$edir', 'elip$zdir'.
Of these, 'elip$center', 'elip$edir', 'elip$zdir', are vectors containing the
longitud, latitud and radial coordinates of the center direction and the two
semiangles, called eta and zeta. 'elip$eta' and 'elip$zeta' are the value of
the two semiangles. All angular magnitudes in radians.
}
\examples{
centerDir <- c(0, pi/2)
axis1Dir <- c(pi/2, 0)
axis2Dir <- c(pi/2, pi/2)
ellipse <- SpherEllipse(centerDir, axis1Dir, axis2Dir, 1/2, 1)
plot(ellipse)
}
\seealso{
\code{\link{plot.SpherEllipse}}
}

