% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SuscTensor}
\alias{SuscTensor}
\title{SuscTensor takes a AMSmeasures object and its setup and returns the
best estimate for the susceptibility tensor as a 3 times 3 matrix.}
\usage{
SuscTensor(measures, setup)
}
\arguments{
\item{measures}{An object of class \code{AMSmeasures}.}

\item{setup}{An object of class \code{AMSsetup}}
}
\description{
SuscTensor takes a AMSmeasures object and its setup and returns the
best estimate for the susceptibility tensor as a 3 times 3 matrix.
}
\examples{
# Experimental setup object
setup <- AMSsetup()
# Sample AMS measures object
measures <- as.AMSmeasures(sample_measures)
# Least squares estimation of the susceptibility tensor
SuscTensor(measures, setup)
}

